[alias]: Remove fasd vim binding; set vim alias
Remove `v` as a fasd binding to open vim using the fasd search as I have
not been using the functionality (and it is a little awkward to use as I
prefer to cd to the directory first).

Set `v` as an alias for `vim` as it is something that I use way more
often.

We can replicate original intended effect by using `j <search>` followed
by `v <Alt-.>`. `Alt-.` adds the last part of previous command (which
would have been expanded by fasd) to the current command.